[#44] fix: 잘못된 좌표계 방위각 문제 수정 (#79)
* [#20] feat: emulator 초기 프로젝트 설정

- emulator 프로젝트 초기 설정을 진행했습니다.
- 스프링에서 사용하는 로깅 설정을 활성화 했습니다 해당 설정은 가져다가 사용시 본인의 입맛에 맞게 수정하시길 바랍니다.
- application 설정 파일을 yml로 변경하고 분할해서 로컬 환경에서 환경변수를 통해 개발자가 서버 정보를 노출하지 않게 했습니다.

* [#20] remove: 중복된 .gitignore 제거

* [#20] chore: 루트 패키지.gitignore 파일 수정

* [#20] feat: emulator 프로젝트 폴더 구조 설정

* [#20] remove: mapper 미사용으로 인한 제거

* [#22] feat: 연동규격서에 따른 도메인 작성

* [#22] feat: 빌더 패턴및 기본값 적용

- 일부 코드라인 수정

* [#22] fix: 잘못된 `@default` 어노테이션을 `@Builder.Default` 으로 변경

- 일부 코드라인 수정

* [#20] feat: emulator 초기 프로젝트 설정

- emulator 프로젝트 초기 설정을 진행했습니다.
- 스프링에서 사용하는 로깅 설정을 활성화 했습니다 해당 설정은 가져다가 사용시 본인의 입맛에 맞게 수정하시길 바랍니다.
- application 설정 파일을 yml로 변경하고 분할해서 로컬 환경에서 환경변수를 통해 개발자가 서버 정보를 노출하지 않게 했습니다.

* [#20] remove: 중복된 .gitignore 제거

* [#20] remove: mapper 미사용으로 인한 제거

* [#22] feat: 연동규격서에 따른 도메인 작성

* [#22] feat: 빌더 패턴및 기본값 적용

- 일부 코드라인 수정

* [#22] fix: 잘못된 `@default` 어노테이션을 `@Builder.Default` 으로 변경

- 일부 코드라인 수정

* [#22] refactor: `@Configuration`을 `@Component`로 수정했습니다.

* [#22] modify: 도메인의 타입을 알맞게 수정

* [#22] feat: 에뮬레이터 상태에 대한 도메인 추가

* [#22] refactor: 도메인 상수 분리및 공통화 적용

* [#44] feat: 스프링 스케줄러 활성화

* [#46] refactor: 도메인 타입 수정

* [#44] feat: 애뮬레이터 관리 컨트롤러 및 페이지 기본 구현

* [#44] modify: db 미사용으로 인한 jpa, mariadb 비활성화

* [#44] remove: 불필요한 파일 삭제

* [#44] feat: 에뮬레이터 활성화및 비활성화 기능 구현

* [#44] feat: 서버 전송을 위한 RestTemplateConfig 생성

* [#44] remove: 컨트롤러 사용으로 인한 제거

* [#44] feat: 시동 ON 정보 전달 구현

- 연동 규격서에 작성된 규칙들은 기본 컨트롤러 구현후 추가할 예정입니다.

* [#44] feat: 서버간의 통신을 위해 RestTemplateConfig 생성 후 설정

* [#44] remove: 미사용으로 인한 클래스 제거

* [#44] feat: 누락된 도메인 속성 수정

- 기능의 테스트를 위해 도메인 클래스별로 toString 메서드를 생성 했습니다.

* [#44] feat: DTO 클래스 생성

- 기능의 테스트를 위해 DTO 클래스별로 toString 메서드를 생성 했습니다.

* [#44] feat: 시동 ON/OFF 및 주기 정보 기능 구현

- 서버 연동 테스트의 경우는 컬렉션 서버가 일부 완성되고 나서 진행 될 예정입니다.

* [#44] refactor: 변경된 컨트롤러 주소를 타임리프에 적용

* [#44] feat: 대시보드의 디자인 개선 및 기능 추가 구현

* [#44] fix: 잘못된 input 필드 값 수정

* [#44] docs: gps 모듈 구현을 위한 gpx 파일 추가

* [#44] feat: gps 모듈 사용을 위한 기능 구현

* [#44] feat: gpx 파일을 랜덤으로 가져오는 기능 구현

* [#44] fix: 불필요한 엘리먼트 요소 제거

* [#44] feat: 404, 500 오류 페이지 적용

* [#44] feat: 루트 주소 리다이렉트 구현

* [#44] chore: jackson 라이브러리 추가

* [#44] fix: jackson 라이브러리 버저닝 오류 수정

* [#44] fix: jackson 라이브러리 버저닝 오류 재수정

* [#44] feat: json 데이터베이스 기능 구현

* [#44] fix: json 데이터가 null로 생성되는 문제를 해결

* [#44] feat: 주기 정보 기능 구현및 고도화

- json db 기능 구현
- gpx 파일을 파싱 구현
- 위도 경도를 기반으로 방향 속도 누적 주행 거리 연산 기능 구현

* [#44] fix: 누락된 도메인 속성 수정

* [#44] feat: 웹 경로 정보 기능 구현

* [#44] feat: 경로 정보 웹 인터페이스 디자인 개선

* [#44] fix: 잘못된 좌표계 방위각 문제 수정

+++ emulator/src/main/java/com/where_car/emulator/gps_module/service/GpsService.java
@@ -42,7 +42,7 @@ public class GpsService {
    * @param lon2 두 번째 지점의 경도
    * @return 방위 (도 단위)
    */
-  public double calculateBearing(double lat1, double lon1, double lat2, double lon2) {
+  public int calculateBearing(double lat1, double lon1, double lat2, double lon2) {
     double longitude1 = Math.toRadians(lon1);
     double longitude2 = Math.toRadians(lon2);
     double latitude1 = Math.toRadians(lat1);
@@ -52,6 +52,6 @@ public class GpsService {
     double y = Math.sin(longDiff) * Math.cos(latitude2);
     double x = Math.cos(latitude1) * Math.sin(latitude2) - Math.sin(latitude1) * Math.cos(latitude2) * Math.cos(longDiff);
 
-    return (Math.toDegrees(Math.atan2(y, x)) + 360) % 360; // 방위 (도 단위)
+    return (int) ((Math.toDegrees(Math.atan2(y, x)) + 360) % 360); // 방위 (도 단위)
   }
 }
